Fix dc_main.cpp: librecomp GCC 9 spaceship and 32-bit hash
Patch librecomp game.hpp/mods.hpp for GCC 9 comparison operators and
32-bit size_t hash packing on SH-4.

+++ src/main/dreamcast/dc_main.cpp
@@ -59,7 +59,7 @@ RspUcodeFunc* get_rsp_microcode(const OSTask* task) {
     case M_NJPEGTASK:
         return njpgdspMain;
     default:
-        fprintf(stderr, "Unknown RSP task: %u\n", task->t.type);
+        fprintf(stderr, "Unknown RSP task: %lu\n", static_cast<unsigned long>(task->t.type));
         return nullptr;
     }
 }
